Put egor wrapping code in a separate module
